Move GameModel away
